functions/requirements.txt
firebase-functions
firebase-admin
google-genai
google-adk
google-cloud-firestore
pydantic
python-dotenv
requests
